requirements.txt
langchain-mistralai
langchain-chroma
langchain-community
langchain-text-splitters
langchain-core
pypdf
python-dotenv
